Refactor: `TabMyTasks` as `UserTasks`
